# tsoa/swagger.json
{
	"components": {
...
		"schemas": {
			"IPost": {
				"properties": {
					"description": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"description",
					"url",
					"createdDate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": ""
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/posts/read/{id}": {
			"get": {
				"operationId": "ReadPostWithId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/posts/create": {
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"createdDate": {
										"type": "string",
										"format": "date-time"
									},
									"url": {
										"type": "string"
									},
									"description": {
										"type": "string"
									}
								},
								"required": [
									"createdDate",
									"url",
									"description"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/posts/update": {
			"put": {
				"operationId": "UpdatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"properties": {
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"id"
										],
										"type": "object"
									},
									{
										"$ref": "#/components/schemas/IPost"
									}
								]
							}
						}
					}
				}
			}
		},
		"/api/posts/delete/{id}": {
			"delete": {
				"operationId": "DeletePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}